hw8/requirements.txt
fastapi==0.115.0
uvicorn==0.32.0
pydantic==2.9.2
asyncpg==0.29.0
psycopg2-binary==2.9.9
numpy==1.26.4
scikit-learn==1.5.2
pytest==8.3.3
pytest-asyncio==0.24.0
httpx==0.27.2
aiokafka==0.11.0
redis==5.0.1
prometheus-client
prometheus-fastapi-instrumentator
sentry-sdk[fastapi]
PyJWT>=2.8.0
